Split shell project unregister tests
